Move face and suit constants to Deck
