+++ lib/texas_holdem/deck.rb
@@ -1,6 +1,31 @@
 require_relative './card'
 
 class Deck
+  class << self
+    def face_values
+      { ' 2' => 2,
+        ' 3' => 3,
+        ' 4' => 4,
+        ' 5' => 5,
+        ' 6' => 6,
+        ' 7' => 7,
+        ' 8' => 8,
+        ' 9' => 9,
+        '10' => 10,
+        ' J' => 11,
+        ' Q' => 12,
+        ' K' => 13,
+        ' A' => 14 }
+    end
+
+    def suit_values
+      { '♣' => 0,
+        '♦' => 1,
+        '♥' => 2,
+        '♠' => 3 }
+    end
+  end
+
   def initialize
     @cards = new_deck
   end
@@ -18,7 +43,7 @@ class Deck
   end
 
   def faces
-    Card.face_values.keys
+    self.class.face_values.keys
   end
 
   def new_deck
@@ -26,6 +51,6 @@ class Deck
   end
 
   def suits
-    Card.suit_values.keys
+    self.class.suit_values.keys
   end
 end
